main 페이지와 검색기능 리팩토링 완료

+++ src/main/java/shop/mtcoding/village/dto/review/response/ReviewList.java
@@ -11,6 +11,6 @@ import lombok.Setter;
 @Setter
 @Getter
 public class ReviewList {
-    private Integer starRating;
+    private Double starRating;
     private Long reviewCount;
 }
